+++ internal/loadproviders/load.go
@@ -7,6 +7,7 @@ import (
 	_ "github.com/wraient/curd/internal/providers/animepahe"
 	_ "github.com/wraient/curd/internal/providers/anineko"
 	_ "github.com/wraient/curd/internal/providers/anipub"
+	_ "github.com/wraient/curd/internal/providers/kickassanime"
 	_ "github.com/wraient/curd/internal/providers/nyaa"
 	_ "github.com/wraient/curd/internal/providers/senshi"
 )